fix(analyze): Tier-2 fixes for member-init, empty-catch, and widening-cast findings
cppcoreguidelines-pro-type-member-init (43 -> 0): real fixes where cheap
(struct stat st{}; POSIX zero-init, ImageRow/MemoryReader given default
member initializers matching their sibling fields' existing style,
ViewPerspective's pos/front/upward given the same default values
resetPos() was setting in the constructor body -- resetPos() itself stays,
since it's still called at runtime to reset the view). NOLINT + comment
for the two cases that can't take a member initializer: the toLt()
conversions in swarm.cpp construct libtorrent system-header types
(lt::dht::public_key/secret_key/signature) that are always fully
overwritten immediately after by a same-size std::ranges::copy.

bugprone-empty-catch (19 -> 0): the 7 in zzstructure_loader.cpp get real
diagnostics.warn() calls -- that function already threads a Diagnostics
object through for exactly this (readColorOr already reports malformed
values the same way), so a malformed spacing/focus_scale/cell_radius/etc.
was being silently ignored with zero feedback. The other 12 are genuine
best-effort fallback-parse patterns (already commented in most cases) that
don't have a diagnostics channel threaded through and don't need one --
NOLINT added since a comment alone doesn't satisfy this check.

bugprone-implicit-widening-of-multiplication-result (168 -> 0) and
bugprone-misplaced-widening-cast (20 -> 0): only ~9 real unique sites
account for both counts (repeated across every TU that includes the
affected headers). Two are genuine integer-overflow risks matching this
session's determinism focus: image_cache.cpp's/image_cache.hpp's
width*height*4 and y*width/y*pitch pixel-offset arithmetic, and
glyphcache/cache.cpp's row*pitch row-offset arithmetic, were all computed
in 32-bit int/unsigned before widening to size_t -- fixed by casting an
operand before the multiply rather than the product after. The rest are
harmless compile-time byte-size constants (16 * 1024 KiB alignments, etc.)
given uz-suffixed literals for the same reason: cheap, and leaves no
NOLINT for a case that isn't actually a live risk.

Verified via full rebuild + all three test suites; the pre-existing,
unrelated ZigzagVisualizerTest.LivingKeymapActionDispatching failure is
untouched by this commit.

+++ include/gleditor/image_cache.hpp
@@ -32,7 +32,8 @@ struct DecodedImage {
 
   [[nodiscard]] bool valid() const {
     return width > 0 && height > 0 &&
-           rgba.size() >= static_cast<std::size_t>(width * height * 4);
+           rgba.size() >= static_cast<std::size_t>(width) *
+                              static_cast<std::size_t>(height) * 4;
   }
 
   [[nodiscard]] float aspectRatio() const {
